feat: Stage 3 - Extract core tools to dedicated module
- Created tools/core.py with 7 core tools from phases 1-3
- Moved: set_slack_token, test_slack_connection, send_message,
  get_channel_info, list_channels, get_user_info, upload_file
- Updated tools/__init__.py to import and register core tools
- Fixed circular import issues with mcp instance
- Removed ~393 lines from main server file
- Updated README.md to document Stage 3 completion

All tools maintain exact signatures and functionality.
