+++ src/slack_mcp/__init__.py
@@ -1,0 +1,6 @@
+"""Slack MCP Server package."""
+
+# The mcp instance will be imported by tools when needed
+# This avoids circular imports
+
+__all__ = []
