+++ src/slack_mcp/tools/__init__.py
@@ -1,0 +1,6 @@
+"""Tool modules for Slack MCP Server."""
+
+# Import all tool modules to ensure registration
+from . import core
+
+__all__ = ["core"]
